+++ doc/examples/api/register_function/register_function_param_zs_int.cpp
@@ -5,7 +5,7 @@ using zetscript::zs_int;
 
 // The C function to register that prints the result of adding (_numbe1) + (*_number2)
 void add(ScriptEngine *_script_engine, zs_int _number1, zs_int  *_number2){
-    printf("Result _number1 + _number2: %lld\n",_number1 + *_number2);
+    printf("Result _number1 + _number2: %ld\n",(long int)_number1 + *_number2);
 }
 
 int main(){
